added clarification
If a user runs the original line that was edited from outside the environment, no error would appear because they would just be creating a new kernel with the same name as "myenv" despite it not actually being "myenv". This could lead to spending some time on debugging, especially for newbies (like me!). So this clarification should prevent that scenario.